+++ examples/akamai_dns_zone/add-zone/dnsv2.tf
@@ -3,7 +3,8 @@ provider "akamai" {
 }
 
 locals {
-  zone = "akavadev.net"
+  pzone = "example_primary.net"
+  szone = "example_secondary.net"
 }
 
 data "akamai_contract" "contract" {
@@ -12,14 +13,26 @@ data "akamai_contract" "contract" {
 data "akamai_group" "group" {
 }
 
-resource "akamai_dns_zone" "test_zone" {
+// Example primary zone resource
+// NOTE: Please review the Provider Getting Started documentation before creating a Primary zone
+//
+resource "akamai_dns_zone" "test_primary_zone" {
 	contract = data.akamai_contract.contract.id
-	zone = local.zone
-	#masters = ["1.2.3.4" , "1.2.3.5"]
+        group = data.akamai_group.group.id
+	zone = local.pzone
 	type = "PRIMARY"
-	comment =  "This is a test zone"
-	group     = data.akamai_group.group.id
+	comment =  "This is a test  primary zone"
 	sign_and_serve = false
+}
+
+resource "akamai_dns_zone" "test_secondary_zone" {
+        contract = data.akamai_contract.contract.id
+        group = data.akamai_group.group.id
+        zone = local.szone
+        masters = ["1.2.3.4" , "1.2.3.5"]
+        type = "secondary"
+        comment =  "This is a test secondary zone"
+        sign_and_serve = false
 }
 
 data "akamai_authorities_set" "ns" {